Make sure to fetch and error about x-readme-* headers from the correct place
Right now this was broken, and we were expecting them to be set on the initial
request object (which would in effect mean from the originating HTTP request).

What we need to do is consume the original request, and make sure the *response*
from the backend API has these headers set. Then use the headers from here as
IDs to send to the metrics backend.

I've tested this using a local instance of hoot.at and ngrok with Cloudflare workers.

https://app.asana.com/0/681252538274980/923335774941624