Extract `HttpRequest` object
`Rack::Request` provides some niceties over a raw env hash but is still
missing some key behavior such as getting HTTP headers or safely reading
the body.

This extracts an object to that wraps `Rack::Request` and provides a
nice interface for these things. Some of the methods just delegate to
existing `Rack::Request` functionality. I made the decision not to try
and isolate the two in testing in order to get more readable tests.