Batch API requests to ReadMe
This commit adds simple request batching behavior to the middleware.

Previously, each request made to an application using this middleware
would cause a new request to be sent to the ReadMe API. A new
`RequestQueue` object was introduced to collect requests and provide a
batched response to send out when ready.

The batch size can be controlled by the user using the `buffer_length`
option. It defaults to 10.