Add mutex to ensure exclusive access when modifying the request queue
This commit adds mutex to `RequestQueue#push`. We found it difficult to
produce a race condition that causes unintended behavior with the
RequestQueue (e.g., multiple requests to the Readme API with the same
data, or requests with empty data), but decided to add the mutex to
ensure such a thing couldn't happen. The mutex ensures that a thread has
exclusive access to the queue so that if the queue is ready, a given
thread will be able to pull off the correct number of payloads from the
queue before releasing access.