Don't submit non-filterable bodies
When given `allow_only` or `reject` parameters, we want to make sure
that the request and response bodies are properly filtered. In order to
do that, we need to be able to parse them. Currently we only support
parsing JSON and form-encoded bodies. If bodies come in as other content
types _and_ a filter is configured, we should not submit the request to
the Readme API because it might contain sensitive data.

If no filter is set, it is safe to just pass-through any bodies to the
Readme API regardless of content type.

Testing this behavior was rather annoying as I had to create new
instances of both app and middleware for each test in order to get
custom response bodies and middleware configurations desired. To make
this nicer I ended up creating a bunch of testing helpers. I also
created two test app classes based on content returned: `JsonApp` and
`TextApp`.

The checking we do around content-types has to happen for both the
request and the response so I pulled it out into a mixin that can be
included into both. Also, as part of this I created an `HttpResponse`
class that decorates `Rack::Request` and includes the mixin.