Don't pass-through requests that fail to parse
While `Metrics` checks that requests have the right content-type for the
filters in place, it is possible for requests to lie about their
content.

If the filter is a "pass-through", then we don't care about the
content type. In fact, we don't want to parse it even if it is JSON
because we'd just be turning it back into a string.

On the other hand, if we can't successfully parse JSON for a
non-"pass-through" filter, we let it raise. The most likely thing that
has happened is that either the request has lied about it having JSON
content or the JSON being submitted is malformed. Either way, this is
catastrophic as we have no way of filtering out sensitive parameters.

`Metrics` has a top-level rescue that will catch this error and cleanly
abort the attempt to log the request to Readme.io.