feat(ruby): accept custom request queue as option
During testing, it can be helpful to see the current user block evaluated
correctly but not actually sent anywhere.

Providing a plain Array as the request queue can be helpful to ensure that
the a request may have been sent.

Providing an object with `push` that is a no-op can be a helpful /dev/null
sink, which permits the middleware to be always present, and not a special
case during testing.