refactor: switch to using ./bin/split.sh from github action
This was quite a bit more involved than I wanted it to be, but hopefully
this will mean we never forget to run it, and it will work consistently.

Tested locally using the following command:

```sh
act -j split-monorepo \
  -s METRICS_SDK_NODE_PRIVATE_KEY="$(<~/.ssh/METRICS_SDK_NODE_PRIVATE_KEY)" \
  -s METRICS_SDK_PHP_PRIVATE_KEY="$(<~/.ssh/METRICS_SDK_PHP_PRIVATE_KEY)" \
  -s METRICS_SDK_PYTHON_PRIVATE_KEY="$(<~/.ssh/METRICS_SDK_PYTHON_PRIVATE_KEY)" \
  -s METRICS_SDK_RUBY_PRIVATE_KEY="$(<~/.ssh/METRICS_SDK_RUBY_PRIVATE_KEY)" \
  -s METRICS_SDK_DOTNET_PRIVATE_KEY="$(<~/.ssh/METRICS_SDK_DOTNET_PRIVATE_KEY)" \
  --container-architecture linux/amd64
```

You'll have to fetch these keys from 1password if you want to run it locally.