fix(dotnet): update SDK to point to /v1/request  (#405)
* fix(dotnet): update SDK to point to /v1/request

This is an updated API endpoint, we're deprecating the version-less one.

* chore(dotnet): add TODO for later when we upgrade deps

Fix version of RestSharp to v1.0.6 for the dotnet v6 example app

This was breaking due to a deprecated interface

* docs(dotnet): add some better example values to the test apps

The lack of a `context.Request.Headers["key"];` when requesting the test
server over GET was causing the requests to not be set. Lets just hardcode
this for the example app.