feat: very first pass at spinning up a mock metrics server to write tests
The theory is that we'll use this, pass in the command to spin up the
child server, and provide the METRICS_SERVER as an env variable that
each of the sdks can look for. Once we can pass that in we can make
HTTP level assertions about what the SDKs can and can't do!

Lets try and write a test first against the node server and see if we
can make it work 🤞