feature: cross language integration tests (dotnet edition) (#435)
* test(integration/dotnet): add super simple dotnet 6.0 test server

Generated using:

```
dotnet new web --no-https
```

* test(integration/dotnet): add very basic .net 6.0 app for metrics

Still outstanding:
- make the sdk respect `METRICS_SERVER` env variable
- try and get `readme__apiKey` better as an env variable
- wire it up to an integration test

* test(integration/dotnet): use regular environment variable name

This is slightly confusing since if they actually wanted to set this up
themselves then they would probably use appsettings.json instead.

This is fine for a demo app though, and will mean it works from the
integration test layer 🤞

* test(integration/node): update tests to work with new port number stuff

* refactor(dotnet): stylistic and whitespace changes

I installed the C# vs code plugin and it did this 🤷‍♂️

* refactor(dotnet): fix warning to do with null variable in test app

* refactor(integration/dotnet): allow the metrics server to be set from the env

Use UriBuilder to construct a URI that always has the correct path on the end

* refactor(dotnet): remove the await from the metrics call so it doesnt block

Right now if metrics is down, this will cause our user's server to stop
responding to requests. This is probably what you'd want in the vast
majority of cases - since it's actually kinda rare you'd wanna fire and
forget an HTTP request.

* chore(integration/dotnet): make host header lookup case insensitive

* test(integration/node): fix up child process usage to use spawn

When using child_process.exec, the process could not reliably be killed
when the test ended, which meant the docker container never closed.

This caused timeouts on the github action:
https://github.com/readmeio/metrics-sdks/runs/6215430863?check_suite_focus=true

* test(integration/node): this no longer needs to be executable

* test(integration): remove the snapshot test for now

There's way too many inconsistencies between even these two web servers
to do jest snapshot testing, so opting to be much more explicit about
assertions for the time being. Would rather get something up and passing
than not running these tests going forward.

* test(integration/dotnet): add docker file for .net test

Added to docker-compose and a new dotnet.yml to be run on gh actions

* refactor: rename incorrect naming of job in gh actions

* test(integration/dotnet): ugh, stupid case insensitive macOS

This failed on github cos it's using a real linux env:

https://github.com/readmeio/metrics-sdks/runs/6226930070?check_suite_focus=true

Lets see if this does the trick 🤞