fix(webhooks): expired signature test
This test was throwing an error, but not the Expired Signature error
that we were expecting. It was throwing because we were passing in
`new Date()` into the test instead of `Date.now()`.

I've fixed all of the tests here, and ensured that we're casting to int
when constructing the new Date() in verify()