fix(python/django): move instantiation of grouping_function
Previously we were trying to import the grouping function in MetricsApiConfig
which in Django's case gets called from settings.py (before the models have
been setup). So if your grouping function was located in a file that happened
to use a model (or was a function on a model) it would error with the following:

```
django.core.exceptions.AppRegistryNotReady: Apps aren't loaded yet.
```

This PR moves the loading of the grouping function (if it needs to be
imported) into Metrics instead of in the Config, this is late enough for
django to have started so you can use models.