fix(node): avoid rethrowing errors when development=true
When development=true, the SDK re-throws in a floating promise, leaving
no way for the caller to catch it. This crashes the node process.

In lieu of proposing an error handling or logging API to use, this
PR/commit is more to point out the problem than suggest a long-term
solve.