refactor(node): moving to native `fetch` (#1053)
## 🐳 Context

The amount of dependencies we use to make a `fetch` call to both the
ReadMe and Metrics APIs is a bit off the rails right now:

* Metrics API requests use` node-fetch` for the POST request to record a
log.
* ReadMe API requests use `api@6` which uses not only `node-fetch`, by
way of `isomorphic-fetch`, but it also uses `oas`, `@readme/oas-to-har`,
`fetch-har`, `json-schema-to-ts`, but also our OpenAPI definition to
make very simple GET requests to retrieve project and version data.
* Even using `api@7` wouldn't be much of a change because that still
uses everything listed her _but_ `node-fetch`. The complexity overhead
for these two small API calls is a bit much. `api@6` also uses a very
old version of our `oas` library and the work involved to upgrade that
here, without just outright moving over to `api@7` is non-trivial. It's
easier to move off it entirely.[^1]

So I'm refactoring all of this away.

## 🧰 Changes

* [x] Moving Metrics API calls from `node-fetch` for native `fetch`.
* We already only support Node 18+ so this shouldn't be considered a
breaking change as `fetch` is readily available to us.
* [x] Completely ditching `api` for ReadMe API calls for native `fetch`.
  * This also allows us to remove our dependency on `json-schema-to-ts`.
* [x] Upgrading `msw` to the latest release as the v2 release better
supports native `fetch`.
* [x] Removes our testing dependency on `form-data` as we have access a
native `FormData` API.[^2]
* Was also able to remove our dependency on `multer` too because in the
tests it was used it wasn't _actually_ being used.
* [x] Removes our dev dependency on `caseless`.
* The thing we were using it for, caseless retrieval of headers in an
object, isn't necessary because HTTP headers always come back to us
lowercased and we were always searching for them lowercased.
* [x] Moved `@types/har-format` to being a dev dependency as we're only
using it for types and aren't exporting anything from it for external
use.

[^1]: There's something to be said for dogfooding our own SDK software
here but because we are unable to do so in any other SDK language, and
`api` is unoptimized due to its design language requiring consulting an
OpenAPI definition before every request IMHO it's not worth it for
something as mission critical as our Metrics SDKs.
[^2]: Not to mention one that actually follows the `FormData` spec.
Don't get me started...

---------

Co-authored-by: Kanad Gupta <8854718+kanadgupta@users.noreply.github.com>