chore(python): Update type hinting on VerifyWebhook's signature (#1032)
The function checks if it's None, so the type hinting should accept None

| 🚥 Resolves ISSUE_ID |
| :------------------- |

I didn't create an issue for this, but I can if that helps.

## 🧰 Changes

Update the type hinting on VerifyWebhook to accurately reflect what the
function can accept. In my application code, I call `signature` with a
variable that can be a `str` or `None`, and mypy is flagging that as a
problem because the function says it only accepts `str`.

## 🧬 QA & Testing

No testing required, since it's just a type hinting change.

I chose to use `typing.Optional[str]` instead of `str | None` because
that works on more versions of python.